+++ Source/Common/Shared/WinCopies.Installer.Shared/Models/ObjectModel/Installer.cs
@@ -12,7 +12,12 @@ namespace WinCopies.Installer
 
         bool InstallForCurrentUserOnly { get; }
 
-        string Location { get; }
+        string
+#if CS8
+            ?
+#endif
+            Location
+        { get; }
 
         string
 #if CS8
@@ -23,17 +28,39 @@ namespace WinCopies.Installer
 
         Actions Actions { get; }
 
+        IEnumerable<KeyValuePair<string, IFile>>
+#if CS8
+            ?
+#endif
+            Files
+        { get; }
+
         bool Completed { get; }
 
         bool RequiresRestart { get; }
 
-        IStartPage StartPage { get; }
+        IStartPage
+#if CS8
+            ?
+#endif
+            StartPage
+        { get; }
 
-        IOptionsData Options { get; }
+        IOptionsData
+#if CS8
+            ?
+#endif
+            Options
+        { get; }
 
         Error Error { get; }
 
-        IInstallerPage Current { get; set; }
+        IInstallerPage
+#if CS8
+            ?
+#endif
+            Current
+        { get; set; }
 
         ExtraData? ExtraData { get; }
     }
@@ -41,6 +68,11 @@ namespace WinCopies.Installer
     public abstract class Installer : IInstaller
     {
         private byte _bools;
+        private IStartPage
+#if CS8
+            ?
+#endif
+            _startPage;
 
         public abstract string ProgramName { get; }
 
@@ -48,7 +80,12 @@ namespace WinCopies.Installer
 
         public bool InstallForCurrentUserOnly { get => GetBit(0); internal set => SetBit(0, value); }
 
-        public string Location { get; internal set; }
+        public string
+#if CS8
+            ?
+#endif
+            Location
+        { get; internal set; }
 
         public string
 #if CS8
@@ -59,15 +96,44 @@ namespace WinCopies.Installer
 
         public Actions Actions { get; internal set; }
 
-        public IEnumerable<KeyValuePair<string, IFile>> Files { get; internal set; }
+        public IEnumerable<KeyValuePair<string, IFile>>
+#if CS8
+            ?
+#endif
+            Files
+        { get; internal set; }
 
         public bool Completed { get => GetBit(1); private set => SetBit(1, value); }
 
-        public IStartPage StartPage { get; }
+        public IStartPage
+#if CS8
+            ?
+#endif
+            StartPage => _startPage
+#if CS8
+            ??=
+#else
+            ?? (_startPage =
+#endif
+            GetStartPage()
+#if !CS8
+                )
+#endif
+            ;
 
-        public IInstallerPage Current { get; set; }
+        public IInstallerPage
+#if CS8
+            ?
+#endif
+            Current
+        { get; set; }
 
-        public IOptionsData Options { get; internal set; }
+        public IOptionsData
+#if CS8
+            ?
+#endif
+            Options
+        { get; internal set; }
 
         public Error Error { get; internal set; }
 
@@ -75,13 +141,17 @@ namespace WinCopies.Installer
 
         ExtraData? IInstaller.ExtraData => null;
 
-        public Installer() => StartPage = GetStartPage();
+        public Installer() { /* Left empty. */ }
 
         private bool GetBit(in byte pos) => _bools.GetBit(pos);
         private void SetBit(in byte pos, in bool value) => UtilHelpers.SetBit(ref _bools, pos, value);
 
         internal void MarkAsCompleted() => Completed = true;
 
-        protected abstract IStartPage GetStartPage();
+        protected abstract IStartPage
+#if CS8
+            ?
+#endif
+            GetStartPage();
     }
 }