+++ Source/Common/Shared/WinCopies.Installer.Shared/ViewModels/Pages/Process/Process.Partial/Process.cs
@@ -2,7 +2,7 @@
 
 namespace WinCopies.Installer
 {
-    public partial class ProcessPageViewModel : CommonPageViewModel2<IProcessPage, IEndPage, IProcessData>, IProcessPage
+    public partial class ProcessPageViewModel : CommonPageViewModel2<IProcessPage, IEndPage, IProcessData>, IProcessPageViewModel
     {
         public sealed override bool CanBrowseBack => false;